extended issue753 to other files
