+++ frontend/src/app/change-password/change-password.component.html
@@ -3,7 +3,7 @@
 
     <h3 translate>TITLE_CHANGE_PASSWORD</h3>
 
-    <div class="form-container">
+    <div class="form-container" id="password-form">
 
       <div class="confirmation" [hidden]="!(confirmation && !passwordControl.dirty && !newPasswordControl.dirty && !repeatNewPasswordControl.dirty)">{{ confirmation }}</div>
