+++ frontend/src/app/complaint/complaint.component.html
@@ -3,7 +3,7 @@
 
     <h3 translate>TITLE_COMPLAIN</h3>
 
-    <div class="form-container">
+    <div class="form-container" id="complaint-form">
 
       <div class="confirmation" [hidden]="!(confirmation && !messageControl.dirty)">{{ confirmation }}</div>
