+++ frontend/src/app/contact/contact.component.html
@@ -9,7 +9,7 @@
       <p class="error">{{error}}</p>
     </div>
 
-    <div class="form-container">
+    <div class="form-container" id="feedback-form">
 
       <input hidden type="text" id="userId" [formControl]="userIdControl"/>
