+++ frontend/src/app/login/login.component.ts
@@ -7,7 +7,7 @@ import { library, dom } from '@fortawesome/fontawesome-svg-core'
 import { UserService } from '../Services/user.service'
 import { faKey, faEye, faEyeSlash } from '@fortawesome/free-solid-svg-icons'
 import { faGoogle } from '@fortawesome/free-brands-svg-icons'
-import { pressEnterHandler } from 'src/main';
+import { pressEnterHandler } from 'src/functions'
 
 library.add(faKey, faEye, faEyeSlash, faGoogle)
 dom.watch()
@@ -67,7 +67,7 @@ export class LoginComponent implements OnInit {
     pressEnterHandler('login-form', () => this.login(), () => this.isEnabled())
   }
 
-  isEnabled() {
+  isEnabled () {
     return this.emailControl.value.trim() !== '' && this.passwordControl.value.trim() !== ''
   }
