+++ frontend/src/app/recycle/recycle.component.html
@@ -4,7 +4,7 @@
 
     <h3 translate>TITLE_RECYCLE</h3>
 
-    <div class="form-container">
+    <div class="form-container" id="recycle-form">
 
       <div class="confirmation" *ngIf="confirmation && !recycleQuantityControl.dirty && !recycleAddressControl.dirty">{{
         confirmation }}
