+++ frontend/src/app/register/register.component.html
@@ -3,7 +3,7 @@
 
     <h3 translate>TITLE_REGISTRATION</h3>
 
-    <div class="form-container">
+    <div class="form-container" id="registration-form">
 
       <mat-form-field appearance="outline">
         <mat-label translate>LABEL_EMAIL</mat-label>
