+++ frontend/src/main.ts
@@ -5,16 +5,6 @@ import { AppModule } from './app/app.module'
 import { environment } from './environments/environment'
 import 'hammerjs'
 
-export function pressEnterHandler (loginForm, loginFunc, isEnabled) {
-  document.getElementById(loginForm)
-    .addEventListener('keyup', function (event) {
-      event.preventDefault()
-      if (event.keyCode === 13 && isEnabled()) {
-        loginFunc()
-      }
-    })
-}
-
 if (environment.production) {
   enableProdMode()
 }
